api: Make WorldPoint extend WorldArea
Functionally, WorldPoints have always behaved like areas with width and
height 1, which is evident by looking at the number of methods between
the two classes which exist to compare distances, convert instances of
one to the other, etc. This commit codifies this extension and removes
now-unneeded API which is now inherited from WorldArea.