api: Remove WorldArea#canTravelInDirection extra condition predicate
This method appears to have been entirely unused since it was originally
added; it was not used in ea5c1f30329535a32c474864788b1dd5fdcdbd3e or
subsequent commits (which were for the old demonic gorilla plugin, [see
pull request 1548][1]), nor has it been used in the time since then.
Given that it is unlikely we will have a need to check NPC collision in
pathfinding, it should be removed.

[1]: https://github.com/runelite/runelite/pull/1548